Improved autocomplete close
The autocomplete close no longer activates after a half-second delay.
This delay was originally used to prevent the box from closing as a user
was clicking an item. Now, a mouseover sets a flag in the classList
which is turned off on mouseout. This flag is prevents the autocomplete
box from closing on clicks, rather than the crude delay that was used
before. Box still closes on item selections and add item popups, but
only after these operations have finished.