Fixed tabbing on autocomplete keyboard controls
Tabbing away from an autocomplete would end up restricted to the form
elements. If the autocomplete was the first element, one could not tab
above it. A relatively minor point, but for accessibilty's sake, this
has been fixed by letting the Tab event's default behaviour run
normally, and only having the simulated behaviour code run when the
Enter key is pressed. Also, moved the preventDefault command to the
locations where it needs to run, rather than a conditional statement
outside it all.