Modified database installation instructions
Installation instructions regarding database setup are more detailed and
remove much ambiguity. Also modified the installation scripts, moving
any permission-related commands to the readme. Finally, made it clear
that the final steps are meant to be run by the admin user, and not the
technician installing the system and running the setup scripts.