Made vacuuming an optional feature.
Vacuuming from the web interface is now disabled by default. It can be
reenabled by altering a constant. This was done because changes to the
setup instructions meant that now the main user account for the web
interface may not be set up as owner of the database tables. Because of
this, they can't run a database vacuum. So I disabled the feature to
reduce confusion. Also altered the help file accordingly.