Fixed major security hole in change email function
A major security hole was found, allowing users to, through a bit of
clever hacking, change any other user's email address to whatever they
wanted. This is now patched to throw an error message if they attempt to
change the email for any user they are not logged in as.