Removed case-sensitivity from login
Login email is now case insensitive. This is mainly to prevent
accidental duplication of email addresses, where the only difference is
one is upper case, and one is lower case. Now, all email addresses are
lower case. This might cause issues for addresses on email servers that
have case-sensitive addresses, but as far as I can tell, while it is
possible, it is not done on any major server.