Added new aspect to add class time
Add class time can now be run without having to search for a CRN in the
course listings. If one knows the CRN of the course they want to add a
time to, they can just add it with no other issue. Also, added an error
catch to the add CRN function, in case an invalid course code is called.