Altered course search
Searching for course codes no longer works for searching from the middle
of a course name. Search terms now must specifiy the initial characters.
This is mainly because I couldn't think of a reason one would search for
course codes by the middle characters.

Also fixed the search by course name aspect, returning the wildcards,
and keeping the ability to search for courses by any part of the course
name.