Removed redundant string functions in courses.php
Removed all calls of the UPPER command in the various SQL prepared
statements. This is because, in every case, when the related PHP
functions are called, the course codes are already uppercase. Either
because the parse_course_code function was called, or the course code
was provided through means other than user input. Therefore, there is no
need to call the UPPER function in the database.