Altered sidebar build function
Sidebar will no longer create invisible links to pages that you
shouldn't be able to access without specifying parameters. This was
originally done for the sake of simplicity, to allow a simple way to
hide these links, while connecting the pages to certain categories in
the sidebar. The pages are still linked to the categories, but they
don't appear in the HTML. This changes nothing with regards to site
appearance, but was merely done because having those useless links was
stupid.